+++ src/index.d.ts
@@ -163,6 +163,26 @@ interface DiscordFeaturesHandlerOptions {
    */
   slashCommandIdsToDelete?: string[] | string;
   /**
+   * If you want to change the behavior to deleting all slash commands when the bot starts up, you can provide an object to define which slash commands to delete
+   *
+   * @default {
+   * delete_global_slash_commands: false,
+   * delete_guild_slash_commands: false
+   * }
+   */
+  onSlashCommandsLoading?: {
+    /**
+     * Delete all global slash commands before loading new ones
+     * @default false
+     */
+    delete_global_slash_commands?: boolean;
+    /**
+     * Delete all guild slash commands before loading new ones
+     * @default false
+     */
+    delete_guild_slash_commands?: boolean;
+  };
+  /**
    * Disable discord-features-handler unhandledRejection handler
    */
   disableUnhandledRejectionHandler?: boolean;
@@ -259,9 +279,10 @@ interface CommandFile {
   usage?: string;
   /**
    * SlashCommandBuilder object for creating this command into a slash command
+   *
+   * @todo: Discord Recommands using slash commands over prefix commands, so this may be required in next version of discord-features-handler
    */
   data?: SlashCommandBuilder;
-
   /**
    * customIds for your interaction components
    * @example
@@ -299,6 +320,9 @@ interface CommandFile {
    * @param interaction Interaction object
    * @param client Client Object
    * @param level permission level of user
+   *
+   * @todo
+   * This function is used for slash commands, context menu commands, and user commands, and is required to be implemented in next version of discord-features-handler, as discord recommends using slash commands over prefix commands; You can still return an empty interaction reply and not implmeent this function and data property.
    */
   interactionReply?(
     interaction: CommandInteraction,
@@ -319,52 +343,6 @@ interface CommandFile {
       | ModalSubmitInteraction
       | UserContextMenuCommandInteraction
       | AutocompleteInteraction,
-    client?: Client,
-    level?: number
-  ): Promise<Interaction>;
-
-  /**
-   * (select or button interaction), if you are using createMessageComponentCollector, you do not need to define this function
-   * Executing a message component interaction command call
-   * @param interaction Interaction object
-   * @param client Client Object
-   * @param level permission level of user
-   */
-  componentInteraction?(
-    interaction: MessageComponentInteraction,
-    client?: Client,
-    level?: number
-  ): Promise<Interaction>;
-  /**
-   * Executing a auto complete interaction command call
-   * @param interaction Interaction object
-   * @param client Client Object
-   * @param level permission level of user
-   */
-  autoCompleteInteraction?(
-    interaction: AutocompleteInteraction,
-    client?: Client,
-    level?: number
-  ): Promise<Interaction>;
-  /**
-   * Executing a modal interaction command call
-   * @param interaction Interaction object
-   * @param client Client Object
-   * @param level permission level of user
-   */
-  modalInteraction?(
-    interaction: ModalSubmitInteraction,
-    client?: Client,
-    level?: number
-  ): Promise<Interaction>;
-  /**
-   * Executing a user context menu interaction command call
-   * @param interaction Interaction object
-   * @param client Client Object
-   * @param level permission level of user
-   */
-  contextMenuInteraction?(
-    interaction: UserContextMenuCommandInteraction,
     client?: Client,
     level?: number
   ): Promise<Interaction>;