+++ docs/release-notes.md
@@ -9,12 +9,23 @@ hide:
 
 # Release Notes
 
-## 2.3.0
-- Added a options to the DiscordFeaturesHandlerOptions to be able to delete specific slash command ids, using slashCommandIdsToDelete 
+## 3.0.0
+
+### Feature
+- Add new optional options to the DiscordFeaturesHandlerOptions
+  - slashCommandIdsToDelete: Array of strings for deleting specific slash command ids,
+  - onSlashCommandsLoading: Object of boolean for enabling delete of slash commands before loading new ones
+- CommandFile Properties update:
+  - Change `customIds` to Array of Strings of customIds or a Object of key values
+  - Removed  `componentInteraction`, `autoCompleteInteraction` `contextMenuInteraction` `modalInteraction` and replaced with a unviersal one called: `customIdInteraction` that handles interaction based off `customIds`
+  - Add `customIdInteraction(interaction, client, level)` to handle customIds interactions to ensure no button or other type of customIds are inactive if the bot application or collector run out on time.
+
+### Fix
+- Issue where prefix commands doens't work
 
 ## 2.2.0
+### Fix
 - Updated event loader to prevent duplicate event listeners.
-- Updated command loader to prevent duplicate slash commands from being created when the bot restarts.
 
 # Release Notes